add new net client
